Move governance module to gov package
